Add eslint and all the things
