--- Aside.js
@@ -1,17 +1,0 @@
-import React from "react"
-
-function Aside() {
-    return(
-        <aside className="sidebar-main">
-            <ul className="sidebar-categories">
-                <li className="sidebar-item"></li>
-                <li className="sidebar-item"></li>
-                <li className="sidebar-item"></li>
-                <li className="sidebar-item"></li>
-                <li className="sidebar-item"></li>
-            </ul>
-        </aside>
-    )
-}
-
-export default Aside
