+++ Aside.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+function Aside() {
+  return (
+    <aside className="sidebar-main">
+      <ul className="sidebar-categories">
+        <li className="sidebar-item" />
+        <li className="sidebar-item" />
+        <li className="sidebar-item" />
+        <li className="sidebar-item" />
+        <li className="sidebar-item" />
+      </ul>
+    </aside>
+  );
+}
+
+export default Aside;
