+++ index.js
@@ -1,17 +1,15 @@
-import React from "react"
-import ReactDOM from "react-dom"
-import NavBar from "./NavBar"
-import ContentContainer from "./ContentContainer"
+import React from "react";
+import ReactDOM from "react-dom";
+import NavBar from "./NavBar";
+import ContentContainer from "./ContentContainer";
 
-class App extends React.Component {
-    render() {
-        return(
-            <div>
-                <NavBar />
-                <ContentContainer />
-            </div>
-        )
-    }
+function App() {
+  return (
+    <div>
+      <NavBar />
+      <ContentContainer />
+    </div>
+  );
 }
 
-ReactDOM.render(<App /> , document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById("root"));
